Updated install script. Updated version to 0.0.2.

+++ install.sh
@@ -4,6 +4,9 @@ IFS=$'\n\t'
 
 PROJECT_NAME="zfind"
 SKIP_CLEANUP=false
+INSTALL=true
+TEST_DEPLOY=false
+PROD_DEPLOY=false
 
 while [[ $# > 0 ]]
 do
@@ -26,6 +29,18 @@ do
         --skip)
             SKIP_CLEANUP=true
             ;;
+        -d)
+            TEST_DEPLOY=true
+            ;;
+        --deploy)
+            TEST_DEPLOY=true
+            ;;
+        -p)
+            PROD_DEPLOY=true
+            ;;
+        --production)
+            PROD_DEPLOY=true
+            ;;
         -v)
             set -x
             ;;
@@ -37,10 +52,25 @@ do
     shift
 done
 
-pip install wheel
-pip list | tr -s ' ' | grep -e "^$PROJECT_NAME " | cut -d' ' -f1 | xargs -I {} pip uninstall -y {} || true
-python3 setup.py bdist_wheel
-ls -1 ./dist | grep -e ".whl$" | xargs -I {} pip install ./dist/{}
+if [ "$INSTALL" == "true" ]; then
+  pip install wheel
+  pip list | tr -s ' ' | grep -e "^$PROJECT_NAME " | cut -d' ' -f1 | xargs -I {} pip uninstall -y {} || true
+  python3 setup.py bdist_wheel
+  ls -1 ./dist | grep -e ".whl$" | xargs -I {} pip install ./dist/{}
+fi
+
+if [ "$TEST_DEPLOY" == "true" ]; then
+  echo "Uploading to test pypi"
+  #twine upload --repository testpypi dist/*
+  pip list | tr -s ' ' | grep -e "^$PROJECT_NAME " | cut -d' ' -f1 | xargs -I {} pip uninstall -y {} || true
+  echo "Reinstalling from test pypi"
+  pip install --index-url https://test.pypi.org/simple/ $PROJECT_NAME
+fi
+
+if [ "$PROD_DEPLOY" == "true" ]; then
+  echo "TODO"
+  exit 1
+fi
 
 if [ "$SKIP_CLEANUP" == "false" ]; then
   rm -rf ./dist && rm -rf ./build && rm -rf "$PROJECT_NAME.egg-info"
